separate TXLink
